+++ frontend/src/pages/DevelopmentDashboard.tsx
@@ -23,6 +23,7 @@ import { CommitHygienePanel } from '../components/dev-dashboard/CommitHygienePan
 import { DocsConsistencyPanel } from '../components/dev-dashboard/DocsConsistencyPanel'
 import { RoadmapDrawer } from '../components/dev-dashboard/RoadmapDrawer'
 import { ReadyStableSection } from '../components/dev-dashboard/ReadyStableSection'
+import { ManualCommandRunsPanel } from '../components/dev-dashboard/ManualCommandRunsPanel'
 import { AIExportPanel } from '../components/dev-dashboard/AIExportPanel'
 import { StandaloneModeBanner } from '../components/dev-dashboard/StandaloneModeBanner'
 import { UpdateStatusCard } from '../components/dev-dashboard/UpdateStatusCard'
@@ -217,6 +218,7 @@ const DevelopmentDashboard: React.FC = () => {
             <>
               <StructuralHealthPanel dashboard={dashboard} t={t} />
               <EvidencePanel dashboard={dashboard} t={t} />
+              <ManualCommandRunsPanel dashboard={dashboard} t={t} />
               <CommitHygienePanel dashboard={dashboard} t={t} />
               <DocsConsistencyPanel dashboard={dashboard} t={t} />
               <AIExportPanel